==> 10-EDA-data.ipynb <==
# 10-EDA-data

> Get some prelimary EDA

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

id_reviews = pd.read_csv('Data/id_reviews.tsv', sep='\t', names = ['id', 'review'])
main = pd.read_csv('Data/main.csv')

id_reviews

id_reviews.at[726, "review"]

main

from sentence_transformers import SentenceTransformer
sentences = ["This is an example sentence", "Each sentence is converted"]

model = SentenceTransformer('sentence-transformers/distiluse-base-multilingual-cased-v1')
embeddings = model.encode(sentences)
print(embeddings)

embeddings.shape
