==> 21-review-description-similarity.ipynb <==
# 21-review-description-similarity

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from sentence_transformers import SentenceTransformer

sentences = ["That is a happy person", "That is a happy dog", "That is a very happy person", "Today is a sunny day"]

model = SentenceTransformer('sentence-transformers/all-MiniLM-L6-v2')
embeddings = model.encode(sentences)
print(embeddings)
print(embeddings.shape)

np.dot(embeddings[0, :], embeddings[1, :])/(np.linalg.norm(embeddings[0, :])*np.linalg.norm(embeddings[1, :]))

np.dot(embeddings[0, :], embeddings[2, :])/(np.linalg.norm(embeddings[0, :])*np.linalg.norm(embeddings[2, :]))

np.dot(embeddings[0, :], embeddings[3, :])/(np.linalg.norm(embeddings[0, :])*np.linalg.norm(embeddings[3, :]))

## Read in main data

main = pd.read_csv('Data/main_211110.csv', dtype=str)
main

## Convert review and description into list for models

reviews = main["review"].tolist()
reviews[:2]

descriptions = main["description"].tolist()
descriptions[:2]

## Create embeddings and calculate cosine similarity scores

model = SentenceTransformer('sentence-transformers/all-MiniLM-L6-v2')
review_embeddings = model.encode(reviews)
print(review_embeddings)
print(review_embeddings.shape)

### calculate three first maximum cosine similarity score movie for each one

mid = main['id'].tolist()
max_movie_id = []
for i in range(review_embeddings.shape[0]):
    cur_id = mid[i]
    cosine = {}
    for j in range(review_embeddings.shape[0]):
        if j != i and mid[j] != cur_id:
            cosine[j]=np.dot(review_embeddings[i, :], review_embeddings[j, :])/(np.linalg.norm(review_embeddings[i, :])*np.linalg.norm(review_embeddings[j, :]))
    sorted_cosine = dict(sorted(cosine.items(), key=lambda item: item[1], reverse = True))
    index = list(sorted_cosine.keys())[:3]
    max_id = [mid[q] for q in index]
    max_movie_id.append(max_id)

max_movie_id
