==> 30-clean-data.ipynb <==
import pandas as pd

# get data
data = pd.read_csv('Data/main.csv')

# drop duplicate columns
data = data[data.columns.drop(list(data.filter(regex='_y')))]

# find missing values of each column
data.isna().sum()

# drop columns that had more than 27% missing values
perc = 27.0
min_count = int(((100-perc)/100)*data.shape[0] + 1)
data = data.dropna(axis=1, thresh=min_count)

# rename columns
data = data.rename(columns={'id_x': 'id', 'title_x': 'title', 'original_title_x':'original_title'})

# delete more duplicate columns
data = data.drop(columns=['production_countries','spoken_languages','title','production_companies','release_date'])

# split each genre into its own columns
data[['genre_1', 'genre_2', 'genre_3']] = data['genre'].str.split(',', expand=True)

# drop genre since they have been split
data = data.drop(columns=['genre'])

# combine both vote columns
data[['votes']] = (data['avg_vote'] + data['vote_average']) / 2

# drop columns for votes
data = data.drop(columns=['avg_vote','vote_average'])

cd Data

data.to_csv('main_211109.csv')
