==> 42-create-kmodes-final-dataframe.ipynb <==
# 42-create-kmodes-final-dataframe
> In this notebook, I am going to create 2 dataframes which will be used for calculating Euclidean distance in the next netebook 50.

There are 4 methods we are planning:

1. including ['votes'] column + Kmode + Kmeans
2. including ['votes'] column + Kmode 
3. NOT including ['votes'] column + Kmode + Kmeans
4. NOT including ['votes'] column + Kmode 

In this notebook, I am going to create 2 and 4.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import os.path
import glob

import warnings
warnings.filterwarnings('ignore')

# Import data

For the first step, we need to import data. If you want to run this code in your local computer, please change the path of "personal_path" corresponding to your local environment.

# Define file paths
prefix_path = os.path.expanduser('~/')
# If you want to run this code, please edit personal path below.
personal_path = 'Vanderbilt University/2021 Fall/'
folder_path = 'case2/Data/cluster16_kmode/'
path = prefix_path + personal_path + folder_path

# Import all csv file in the cluster16_kmode folder.
csv_files = glob.glob(path + "*.csv")

Let's import all csv files and save this in "df_list". By doing this, we can call every csv files quickly and it helps us handle many data easily.

# Import dataframe and assign them to a list
df_list = []
for i in range(len(csv_files)):
    df_list.append(pd.read_csv(csv_files[i]))

let's create a new column "group". This value is refers to a group number after we diveded the data using Kmode in the previous notebook 40.

for i in range(len(df_list)):
    df_list[i]['group'] = i + 1

Now, let's combine all dataframe. The reason why we are combining all data together is because we are going to normalize data. 

# Merging multiple dataframes
df_total = [df_list[0], df_list[1],df_list[2], df_list[3], df_list[4], df_list[5], df_list[6], df_list[7], df_list[8], df_list[10], df_list[11], df_list[12]]
all_df = pd.concat(df_total)

# Pre-processing dataframe

When we clustering, we do not need 'recommendation1','recommendation2','recommendation3','recommendation4','recommendation5','recommendation6' columns. Let's remove these 6 columns.

all_df = all_df.drop(['recommendation1','recommendation2','recommendation3','recommendation4','recommendation5','recommendation6'], axis = 1)

Also, in the previous notebook 41, we found that there were 14 movies having no values. Let's remove this movies as well.

no_values_id = ['tt0246425',
 'tt0442781',
 'tt0043852',
 'tt0906783',
 'tt0103790',
 'tt0370919',
 'tt0238119',
 'tt0033368',
 'tt0100100',
 'tt0253614',
 'tt0142634',
 'tt0082892',
 'tt0131636',
 'tt0276868']

final_df_temp = all_df[~all_df["id"].isin(no_values_id)]

By not using some groups, the group's number is messed. Let's reorganize this.

final_df_temp["group"].unique()

final_df_temp.loc[final_df_temp["group"] == 11, "group"] = 10
final_df_temp.loc[final_df_temp["group"] == 12, "group"] = 11
final_df_temp.loc[final_df_temp["group"] == 13, "group"] = 12

final_df_temp.head()

# Noramlization

Similarity is based on the distance between data point's features. For this reason, normalization is very important for calculating distance. I am going to use min-max normalization to each column. After this process, values will be between 0 to 1.

def normalize(df):
    """
    Normalize all values in dataframe
    
    Parameters
    ----------
    df : dataframe
        a dataframe we want to normalize
        
    Returns
    -------
    df 
        a normalized dataframe
    """
    result = df.copy()
    for feature_name in df.columns:
        max_value = df[feature_name].max()
        min_value = df[feature_name].min()
        result[feature_name] = (df[feature_name] - min_value) / (max_value - min_value)
    return result

final_df_temp_id = final_df_temp[['id', 'group']]
left_over = final_df_temp.drop(['id', 'group'], axis = 1)

# Perform normalization
df_norm = normalize(left_over)

df_norm.head()

df_final = pd.concat([final_df_temp_id, df_norm], axis = 1)

df_final.head()

Great! We just created "2. including ['votes'] column + Kmode" dataframe.

To create "4. NOT including ['votes'] column + Kmode" dataframe, let's remove 'votes' column.

df_final_no_votes = df_final.drop(['votes'], axis = 1)

df_final_no_votes.head()

# Save each dataframe to csv files

df_final.to_csv(prefix_path + personal_path + 'case2/Data/yesVotes_kmode_final_df.csv')

df_final_no_votes.to_csv(prefix_path + personal_path + 'case2/Data/noVotes_kmode_final_df.csv')
