==> 50-get-5recommendation-using-euclidian-distance.ipynb <==
# 50-get-5recommendation-using-euclidean-distance
> In this notebook, I am going to get the 5 recommendation movies using Euclidean distance.

Now, we have 4 datasets with different clustering methods and different columns.

1. including ['votes'] column + Kmode + Kmeans
2. including ['votes'] column + Kmode 
3. NOT including ['votes'] column + Kmode + Kmeans
4. NOT including ['votes'] column + Kmode 

Using Euclidian distance, let's create a function returning the 5 closest movie id.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import os.path
import glob

from scipy.spatial import distance

import warnings
warnings.filterwarnings('ignore')

# Import data

# Define file paths
prefix_path = os.path.expanduser('~/')
# If you want to run this code, please edit personal path below.
personal_path = 'Vanderbilt University/2021 Fall/'
folder_path = 'case2/Data/'

kmodes_withVotes = 'yesVotes_kmode_final_df.csv'
kmodes_noVotes = 'noVotes_kmode_final_df.csv'
kmodes_kmeans_withVotes = 'yesVotes_kmode_kmeans_final_df.csv'

# 1. including ['votes'] column + Kmode + Kmeans
df_kmodes_withVotes = pd.read_csv(prefix_path + personal_path + folder_path + kmodes_withVotes)

# 2. including ['votes'] column + Kmode 
df_kmodes_kmeans_withVotes = pd.read_csv(prefix_path + personal_path + folder_path + kmodes_kmeans_withVotes)

# 4. NOT including ['votes'] column + Kmode 
df_kmodes_noVotes = pd.read_csv(prefix_path + personal_path + folder_path + kmodes_noVotes)

####### at this time, Jay is working on creating "3. NOT including ['votes'] column + Kmode + Kmeans" data#########
####### Once she finished her job, we can also get 3rd way's recommendation #################

# Drop unnecessary column 'Unnamed: 0'
df_kmodes_withVotes = df_kmodes_withVotes.drop(['Unnamed: 0'], axis = 1)
df_kmodes_noVotes = df_kmodes_noVotes.drop(['Unnamed: 0'], axis = 1)
df_kmodes_kmeans_withVotes = df_kmodes_kmeans_withVotes.drop(['Unnamed: 0'], axis = 1)

df_kmodes_withVotes.shape

df_kmodes_withVotes.head()

# Create a function to recommend the 5 closest movie ids when only used Kmode

def get_5recommendation_noKmeans(df, movie_id):
    """
    Get the dataframe having distance between input movies and others, and 5 recommendation movie ID
    (This function only treats a dataframe created by NO K-means clustering)
    
    Parameters
    ----------
    df : dataframe
        a dataframe among 4 I mentioned above
    movie_id : str
        input movie's id
        
    Returns
    -------
    group_df
        a dataframe having only movies within the same group as input movie + distance between input movie to others
    recommendation
        a list containig 5 movie's id (The 5 closest distance movie's id)
    """
    # Find the group number of input movie.
    group_num = df[df["id"] == movie_id]['group'].values[0]
    # Get dataframe of which the group number is same as the input movie's group
    group_df = df[df["group"] == group_num].reset_index(drop = True)
    # Convert input movie's features into list
    standard_point = group_df[group_df["id"] == movie_id].loc[:, "review_1st":"popularity"].values.tolist()
    
    # Calculate the euclidean distance between input movie and other movies within the same cluster
    for i in range(group_df.shape[0]):
        point = group_df.iloc[i]["review_1st":"popularity"].values.tolist()
        
        group_df.loc[i, "distance"] = distance.euclidean(standard_point, point)
    
    group_df_temp = group_df[group_df.id != movie_id]
    recommendation = group_df_temp.sort_values(by = ['distance'])[:5]['id'].values.tolist()
    
    return group_df, recommendation

# Create a function to recommend the 5 closest movie ids when used both Kmode and Kmeans

def get_5recommendation_yesKmeans(df, movie_id):
    """
    Get the dataframe having distance between input movies and others, and 5 recommendation movie ID
    (This function only treats a dataframe created by K-mode AND K-means clustering)
    
    Parameters
    ----------
    df : dataframe
        a dataframe among 4 I mentioned above
    movie_id : str
        input movie's id
        
    Returns
    -------
    subgroup_df
        a dataframe having only movies within the same group as input movie + distance between input movie to others
    recommendation
        a list containig 5 movie's id (The 5 closest distance movie's id)
    """
    # Find the group number of input movie.
    group_num = df[df["id"] == movie_id]['group'].values[0]
    # Find the subgroup number of input movie.
    subgroup_num = df[df["id"] == movie_id]['subgroup'].values[0]
    # Get dataframe of which the group number is same as the input movie's group
    subgroup_df = df[(df["group"] == group_num)&(df["subgroup"] == subgroup_num)].reset_index(drop = True)
    # Convert input movie's features into list
    standard_point = subgroup_df[subgroup_df["id"] == movie_id].loc[:, "review_1st":"popularity"].values.tolist()
    
    # Calculate the euclidean distance between input movie and other movies within the same cluster
    for i in range(subgroup_df.shape[0]):
        point = subgroup_df.iloc[i]["review_1st":"popularity"].values.tolist()
        
        subgroup_df.loc[i, "distance"] = distance.euclidean(standard_point, point)
    
    subgroup_df_temp = subgroup_df[subgroup_df.id != movie_id]
    recommendation = subgroup_df_temp.sort_values(by = ['distance'])[:5]['id'].values.tolist()
    
    return subgroup_df, recommendation

# Test

Let's test our functions.

movie_id = "tt0096486"

Let's remind the 4 recommendation methods:

1. including ['votes'] column + Kmode + Kmeans
2. including ['votes'] column + Kmode 
3. NOT including ['votes'] column + Kmode + Kmeans
4. NOT including ['votes'] column + Kmode 

# Get recommendation using "1. including ['votes'] column + Kmode + Kmeans"
_, recommendation5_yesKmeans_noVotes = get_5recommendation_yesKmeans(df_kmodes_kmeans_withVotes, movie_id)

# Get recommendation using "2. including ['votes'] column + Kmode" way
_, recommendation5_noKmeans_yesVotes = get_5recommendation_noKmeans(df_kmodes_withVotes, movie_id)

# Get recommendation using "4. NOT including ['votes'] column + Kmode"
_, recommendation5_noKmeans_noVotes = get_5recommendation_noKmeans(df_kmodes_withVotes, movie_id)

recommendation5_yesKmeans_noVotes

recommendation5_noKmeans_yesVotes

recommendation5_noKmeans_noVotes

2 and 4 shows exactly same recommendation movie IDs. When we look at 1's recommendation list, the first 2 movie ids are same as 2 and 4, but the last 3 movie ids are different.

Since the data is normalized, actually one feature 'Votes' cannot give a big impact on the distance. For this reason, I think with "votes" VS without "votes" actually does not matter in our modeling.
