==> 51-yes-votes-evaluation.ipynb <==
# 51-yes-votes-evaluation

#### This notebook will be used to evaluate the models used on datasets that contain user ratings.

### Install packages & run necessary notebooks

# import packages
import pandas as pd
import numpy as np

# run notebook to get pre-defined functions
%%capture
%run 50-get-5recommendation-using-euclidian-distance.ipynb

### Read in datasets

# read in data

path = 'data/'

kmodes = 'yesVotes_kmode_final_df.csv'
kmodes_kmeans = 'yesVotes_kmode_kmeans_final_df.csv'


kmodes_data = pd.read_csv(path + kmodes)
kmodesKmeans_data = pd.read_csv(path + kmodes_kmeans)

imdb_recs = 'imdb_recommendations_tid.csv'
imdb_data = pd.read_csv(path + imdb_recs)

## MODEL 1: KMODES ONLY

### Evaluating for one movie id as test

# test movie id
test = 'tt0374180'

# run model to get predictions
_, test1 = get_5recommendation_noKmeans(kmodes_data, test)
test1

# look up movie id in the imdb recommendations dataset
test2 = imdb_data.loc[imdb_data['id'] == test]
test2

The predictions from our model are not listed in the IMDB recommendations at all. This happened for several movies that I checked.

### Run predictions on full dataset (3k movie ids) and evaluate

# run function to get predictions for all movies - store as dictionary
recs = {}
kmodes_ids = kmodes_data['id'].tolist()
for i in kmodes_ids:
    _, recs_5 = get_5recommendation_noKmeans(kmodes_data, i)
    recs[i] = recs_5

# make dataframe with preds as a list
recs_data = pd.DataFrame(list(recs.items()),columns = ['id','recommendations'])

# convert all 12 recommendations to list to match preds df
cols = list(imdb_data.filter(regex='recommendation').columns)
imdb_data['imdb_recs'] = imdb_data[cols].values.tolist()

# merge imdb preds and model preds
compare_df = recs_data.merge(imdb_data[['id','imdb_recs']], how='inner',on='id')

# for each movie id compare the two lists to see how many are accurate
for i in range(len(compare_df)):
    recs_sum = sum(a == b for a, b in zip(compare_df['recommendations'][i], compare_df['imdb_recs'][i]))
    compare_df['acc'] = recs_sum / 5

Accuracy is 0.0 for all movies ids.

## MODEL 2: KMODES + KMEANS

### Evaluating for one movie id as test

# test movie id
test = 'tt0077631'

# run model on test movie id
_, test1 = get_5recommendation_yesKmeans(kmodesKmeans_data, test)
test1

# filter imdb dataset for test movie id
test2 = imdb_data.loc[imdb_data['id'] == test]
test2

Same as above. No predictions found in IMDB recommendations.

### Run model on full dataset and evaluate

# run function to get predictions for all movies - store as dictionary
recs_kmeans = {}
kmodesKmeans_ids = kmodesKmeans_data['id'].tolist()
for i in kmodesKmeans_ids:
    _, recsKmeans_5 = get_5recommendation_yesKmeans(kmodesKmeans_data, i)
    recs_kmeans[i] = recsKmeans_5

# make dataframe with preds as a list
recsKmeans_data = pd.DataFrame(list(recs.items()),columns = ['id','recommendations'])

# merge imdb preds and model preds
compareKmeans_df = recsKmeans_data.merge(imdb_data[['id','imdb_recs']], how='inner',on='id')

# for each movie id compare the two lists to see how many are accurate
for i in range(len(compareKmeans_df)):
    recsKmeans_sum = sum(a == b for a, b in zip(compareKmeans_df['recommendations'][i], compareKmeans_df['imdb_recs'][i]))
    compareKmeans_df['acc'] = recsKmeans_sum / 5

Accuracy is also 0.0 for all movie ids.
