==> 52-no-votes-evaluation.ipynb <==
# 52-no-votes-evaluation

#### This notebook will be used to evaluate the models used on datasets WITHOUT user ratings

### Install packages and run necessary notebooks

# import packages
import pandas as pd
import numpy as np

# run notebook to get pre-defined functions
%run 50-get-5recommendation-using-euclidian-distance.ipynb

### Read in datasets

# read in data

path = 'data/'

kmodes = 'noVotes_kmode_final_df.csv'
kmodes_kmeans = 'noVotes_kmode_kmeans_final_df.csv'

kmodes_data = pd.read_csv(path + kmodes)
kmodesKmeans_data = pd.read_csv(path + kmodes_kmeans)

# get votes data
votes_data = pd.read_csv(path + 'model_data.csv')

## MODEL 1: KMODES ONLY

### Evaluating for one movie id as test

# test movie idea
test = 'tt0374180'

# run model to get predictions
_, test1 = get_5recommendation_noKmeans(kmodes_data, test)
test1

# get average user ratings for the predictions
ratings = []
for i in test1:
    curr_rate = votes_data.loc[votes_data['id'] == i, 'votes'].iloc[0]
    ratings.append(curr_rate)
    avg_ratings = sum(ratings)/len(ratings)

# print average ratings
avg_ratings

### Run predictions on full dataset (3k movie ids) and evaluate

# run function to get predictions for all movies - store as dictionary
recs = {}
kmodes_ids = kmodes_data['id'].tolist()
for i in kmodes_ids:
    _, recs_5 = get_5recommendation_noKmeans(kmodes_data, i)
    recs[i] = recs_5

# make dataframe with preds as a list
recs_data = pd.DataFrame(list(recs.items()),columns = ['id','recommendations'])

# for each recommendation, get the rating then average them
avg_votes = []
for i in range(len(recs_data)):
    recs_list = recs_data['recommendations'][i]
    ratings = []
    for j in recs_list:
        curr_rate = votes_data.loc[votes_data['id'] == j, 'votes'].iloc[0]
        ratings.append(curr_rate)
        avg_ratings = sum(ratings)/len(ratings)
    avg_votes.append(avg_ratings)

# get the average for the full dataset - used as evaluation for the whole model
sum(avg_votes)/len(avg_votes)

## MODEL 2: KMODES + KMEANS

### Evaluating for one movie id as test

# get test id
test = 'tt0374180'

# run model on test movide id
_, test1 = get_5recommendation_yesKmeans(kmodesKmeans_data, test)
test1

### Run predictions on full dataset (3k movie ids) and evaluate

# run function to get predictions for all movies - store as dictionary
recs_kmeans = {}
kmodesKmeans_ids = kmodesKmeans_data['id'].tolist()
for i in kmodesKmeans_ids:
    _, recsKmeans_5 = get_5recommendation_yesKmeans(kmodesKmeans_data, i)
    recs_kmeans[i] = recsKmeans_5

# make dataframe with preds as a list
recsKmeans_data = pd.DataFrame(list(recs.items()),columns = ['id','recommendations'])

# for each recommendation, get the rating
kk_avg_votes = []
for i in range(len(recsKmeans_data)):
    kk_recs_list = recsKmeans_data['recommendations'][i]
    kk_ratings = []
    for j in recs_list:
        kk_curr_rate = votes_data.loc[votes_data['id'] == j, 'votes'].iloc[0]
        kk_ratings.append(kk_curr_rate)
        kk_avg_ratings = sum(kk_ratings)/len(kk_ratings)
    kk_avg_votes.append(kk_avg_ratings)

# get the average of the ratings for full dataset
# evaluation for the model
sum(kk_avg_votes)/len(kk_avg_votes)
